Move TcpPortOptions and SerialPortOptions into a submodule
